Move implementation to separate file and address PR comments
